Open files in file-tree
Summary:
This diff adds a feature that shows the list of currently open files on top of the file-tree.
The list allows switching between the open files and closing them.
It also shows the `modified` status of the files, and could be seen as a possible replacement for the built-in `tabs` package.

If not desired, the feature can be disabled from the file-tree settings section.

Reviewed By: ssorallen

Differential Revision: D3345006

fbshipit-source-id: b3eb15a4ff23fe5755cccc57d566fa6a520a1f25