Added ability to dismiss (end) a NuxTour
Summary:
Added the ability to end/dismiss a NuxTour. Doing this will count the tour as completed, but send a message (track) that it was skipped.

An "X" has been added to the top right of the Nux and is used to trigger the disposal/forceful ending of the whole tour.

Reviewed By: ssorallen

Differential Revision: D3410343

fbshipit-source-id: 6b4b1cf52644396abd250613a35edfac87ad1d53