Python: Follow imports when hyperclicking if possible
Summary: Commonly, the closest assignment of a name is at an import statement. This is rarely ideal, requiring the user to perform a second hyperclick on the import statement to find the actual definition. Change get_definitions to follow imports until a non-import definition is found, or until it reaches a deadend (i.e. an unresolvable import).

Reviewed By: hansonw

Differential Revision: D3415587

fbshipit-source-id: 51bc16fe067067956616675dcd530245501d327c