Clang: implement a server state
Summary:
This takes a page from the Flow book and implements a server status for the Clang server. Currently there are only two states (ready and compiling).

A wrapper `compile` function in `ClangServer` records the number of active compilation commands and updates the state accordingly. Most commands instantly return if the server is compiling, except for the outline view.

This is easier to work with using the service-framework based RpcProcess.

Reviewed By: nmote

Differential Revision: D3397069

fbshipit-source-id: 6a066f052c7217dcb9c37be5f966bc0fced0e77d