Changed one of NuxManager's internal containers to a Map
Summary:
Use a Map to store Nux information inside of the `NuxManager` for `_pendingNuxes`. Each NUX having a unique ID allows for easier lookup and deletion.

This is not done for the `_readyToDisplay` nux list, since  I want order to matter here. While the values in maps can be accessed in terms of insertion order, this is not ideal since while display order currently corresponds with insertion order into this array, it may not in the future (this may be implemented as a priority queue).

Reviewed By: jgebhardt

Differential Revision: D3402368

fbshipit-source-id: 9ffca82e9bac1e856d5f933dd6146bec66112ea4