Fix process stream completion
Summary:
While working on a feature, I noticed that stream errors weren't actually completing the stream. This fixes that and also makes some of the logic more obvious/explicit.

I added a test which illustrates the problem. (It fails before this commit.)

I also removed an unnecessary race in observeStream.

Differential Revision: D3416432

fbshipit-source-id: 6a0638bf0627a87953fdcece1d78a64f90eced29