Clear filter on file tree and sidebar toggle with integration test
Summary:
The file tree's filter doesn't clear when the sidebar or file-tree is toggled, which makes for bad UX when you enter a filter, toggle off then on the bar, try typing another filter -- you get `filter1filter2` as the result.

This diff adds this feature by clearing the filter when both the `nuclide-file-tree:toggle` and `nuclide-side-bar:toggle` commands are called, along with a spec.

Differential Revision: D3417164

fbshipit-source-id: 2295e30a7781b753ef61cea86b609b7cfecd0fa9