Add lint rule to prevent path imports
Summary:
This lint rule prevents accidental import of the built-in `path` module.
It does not prevent `require`-ing of the module, as `require` is most often used in a non-transpiled code, which is either a script or is part of the transpiler itself.
In both of the above cases we do want to have access to the actual `path`

Differential Revision: D3414273

fbshipit-source-id: ac09c00f6a65491d3fa63933eab50bdcac9bf3a6