Share service config loader between the server and the transpiler
Summary:
This diff makes it possible for the server (or local rpc client) and the transpiler to use the same utility to get the services config.

Service config files must refer to services either with an absolute path or a path relative to the config file. The various loading/reading functions from `nuclide-rpc` that interact with service paths, only accept absolute paths. This reduces the number of possible file resolution site to just one: `loadServicesConfig.js`.

This change is groundwork for generating the rpc proxies at build time.

Reviewed By: peterhal

Differential Revision: D3416184

fbshipit-source-id: 240a60fa416c9e65789b2ca69fe0f83ed8a15ab7