Added limit to number of NUXes that can be displayed per session
Summary:
Added a limit to the number of NUXes that can be displayed every session/window that opens.

Every time a new Nuclide Window opens, the packages are re-initialized, so a new set of `NUX_PER_SESSION_LIMIT` NUXes can be shown. However, since `tryTriggerNux` is hardcoded and will try to trigger a NUX that may already have been shown, it may throw an error on subsequent package loads, since if the NUX is already seen and failed to register it doesn't exist in the NUX framework's internal list when the trigger attempt is made.

So in usage, all `tryTriggerNux`es should be wrapped with a `try{ ... } catch (e) { ... }` block for now. This will be fixed in a later (stacked) diff.

Reviewed By: jgebhardt

Differential Revision: D3428152

fbshipit-source-id: da79654e6de78b65463489d594cf94d21bcc896e