Hook FlowVersion into FlowRoot
Summary:
This allows methods in FlowRoot to learn what the current Flow version is.

The version is invalidated whenever the server state changes to "not running"
since that indicates that the previous server has died for some reason. However,
it is possible that the Flow version will change without us noticing that the
server has died (if it has been started from outside of Nuclide), so the version
is also invalidated when it is 10 minutes old so we are not permanently stuck
thinking the version is wrong (this logic is built into the FlowVersion class).

Reviewed By: peterhal

Differential Revision: D3433135

fbshipit-source-id: f9c938806326a41db2c731579eb538eb3c7bab9e