Dismiss NUX if the element its bound to is hidden
Summary:
Currently if the NUX is bound to an element and the element is hidden/closed (e.g. the Outline View panel), the NUX persists. While the diff before this one in the stack added a "dismiss" button, the floating tooltip is just BAD UX. Time for a fix!

The DOM is polled at a given frequency, and the NUX is dismissed if the element is hidden.
Determining whether the element is hidden is done by checking the `offsetParent` property of its parent, or by using the computed style (much slower, and avoided when possible).

NOTE: Also added minor comment punctuation and grammar changes for consistency throughout the file.

Reviewed By: jgebhardt

Differential Revision: D3419845

fbshipit-source-id: b40d461b0d527f84cb725529c0dd2791153bd730