Explain why .hg/hgrc is not guaranteed.
Summary:
I was looking at an Eden revision where we stopped testing for `.hg/hgrc` in
favor of `.hg/store` because `.hg/hgrc` is not guaranteed to be present.
That reminded me of https://github.com/facebook/nuclide/issues/345, so I
grepped for `hgrc` in Nuclide to see if we were making the same mistake.

It looks like mostafaeweda fixed this in early February:
https://github.com/facebook/nuclide/commit/13305bd2222e13c479091eb2fcd3b012cfddda35

In the end, I didn't have much to add, but I decided to do a little cleanup
while here.

Reviewed By: mostafaeweda

Differential Revision: D3450563

fbshipit-source-id: dc848199441d0213f4851d8deb998374fa851f07