Log and track error in NuxView displaying instead of throwing
Summary:
The `_createNux` method inside the NUX may get called at a later time due to a `setTimeout` call. If this is the case, then we may (and usually do) hit a weird case, where the function to execute (`clearNux`) is called inside of its parent's scope and thus has access to the parent's scope variables, but the parent scope's execution may be over. If this is the case, then there will be nothing to catch the error that is thrown, and will result in a nasty Nuclide error being shown in the top right.

So, the error has been converted to a message that is logged to the console, along with a tracking event sent.

Reviewed By: ssorallen

Differential Revision: D3421639

fbshipit-source-id: dc05b888088b0676e8a589f57b34e7f8bad36193