Diff View: Refactor revisions state fetching and caching without promise chaining
Summary:
This makes it much simpler to change the revisions state logic without hacking around the promise-chain.
Before this, we were fetching the latest `RevisionsState` - then amend it by the currently selected commit id.
This doesn't scale (complexity-wise) when we have more state to amend.

Here, we remove that promise-chaining technique in favor of a more natural caching for only the fetched data, while the `RevisionsState` is constructed when needed.

Reviewed By: ssorallen

Differential Revision: D3414036

fbshipit-source-id: 4ef286e3d5725dd8710f1bd36ce1b7984b86aaf6