Add defs to proxy factories and always use them
Summary:
Embed the defs right in the proxy module. In the case where you generate the proxies at build time, this is way easier than having a separate file for them. It also saves you the hassle of having to deserialize JSON into the definition format, since the defs saved in the exact form that they will be consumed by the service registry.

Had to bump diff views timeout. I think that reading the fat proxy files was too much. In development, this change makes things a bit slower because you have to generate the defs, then re-read them from the proxy module. In builds though, it's crazy fast because there are no defs to generate!

Ideally you'd cache the built proxy modules in development, however coming up with a cache key is tricky. There's no simple way to version the parsing/generation because it's split across multiple files. And there's no simple way to version the input because some service files reference types files.

Reviewed By: peterhal

Differential Revision: D3424048

fbshipit-source-id: d5974c13ccee9a8e679c77832196a096edc3b2ec