Add option to transpile and generate RPC proxies before running tests
Summary: This brings the test environment one step closer to being like production releases. Passing `--destructive-compile` will run the tests on the same code that users will run. It's behind a flag because the process is destructive: the transpile will overwrite the original source, and the proxy generation will create untracked files. This is something you may only want to do during CI.

Reviewed By: peterhal

Differential Revision: D3454739

fbshipit-source-id: 10ccbd3f45175efa1ad9e78138eb9bd1c704aa64