Fix nuclide-settings' TestUtils import and key warning
Summary:
Requiring `react-addons-test-utils` is incorrect. It's a transitive dependency, not a direct dependency. It's only worked so far because we dedupe dependencies, and coincidentally `react-addons-test-utils` is brought top-level. The correct way to import it is through `react-for-atom`.

Also fixes:

```
Warning: Each child in an array or iterator should have a unique "key" prop. Check the render method of `SettingsSelect`. See https://fb.me/react-warning-keys for more information.
```

Reviewed By: bottledwalter

Differential Revision: D3471869

fbshipit-source-id: c794eff6229f78060baf174bffa8452d69ce00e1