Hotfix context view never unsubscribing to definition service
Summary: This is a hotfix for the remote hack service timing out when querying for a definition (T11855776). Context view was greedily subscribing to the definition service, even when the panel was closed. Now, context view subscribes only when (a) there is a definition service to subscribe to and (b) the panel is showing.

Reviewed By: nmote

Differential Revision: D3473999

fbshipit-source-id: 9ee3868831d0dda1f44799d44191bc625bcdec20