Add type for ObservableDiagnosticUpdater [7/n]
Summary:
Largely mimics the previous API, but uses Observables rather than callbacks.

fileMessageUpdates is still a function. The alternative would be to have it be
an Observable that includes all file changes, and allow the client to filter
themselves. But, to satisfy the contract that initial values are provided on
subscribe, we would have to compute and send over all file messages, which could
be expensive. This isn't just academic because we currently subscribe to file
messages every time a new editor is opened.

Reviewed By: matthewwithanm

Differential Revision: D3472370

fbshipit-source-id: 690cd0996a7c8bac2b1bff33bbd889b6f307902f