Fix opening and saving remote files that don't exist yet
Summary:
This fixes two issues: (1) The red box with `"Uncaught Error: getDigestSync is not supported in RemoteFile; name: Error; message: getDigestSync is not supported in RemoteFile"`, and (2) corrects the behavior of `isModified` for buffers with remote files that didn't exist.

The `getDigestSync` error is the result of Atom trying to periodically serialize its state, and the remote file not being able to handle `getDigestSync` when it's never been read. The fix is to implement `getDigestSync` to be behave how Atom's behaves for files that don't exist.

Reviewed By: ssorallen

Differential Revision: D3480338

fbshipit-source-id: 967a25f0e13409b8da6a1cbd9ccdebee69dddd7d