Console diff #4 -- Add "Cached Mode" to LazyNestedValueComponent
Summary:
The semantics for console evaluation are different from the watch window, because we don't want to refetch values from the backend once we already have them.  An example of where this is useful is the following scenario:

1. Attach debugger and evaluate things in the console.
2. Detach the debugger and try to expand and collapse the things you evaluated.

Since the backend has gone away, if we don't do any caching, then we get a loading spinner that will never return.

This caching is also useful because people don't expect their (already expanded) expressions to re-evaluate if some later side-effect occurs.

The approach I took here is to:
1. Only cache ExpansionResults gotten from the backend, and
2. Store these cached ERs in the TopLevelComponent, so that this data can persist across re-renders.

Reviewed By: jgebhardt

Differential Revision: D3472324

fbshipit-source-id: 95b52833d7aa3ab25cdd8e16d2e2e32c891c3914