Gracefully handle definition service errors, reduce frequency of definitions being emitted from definition service subscription
Summary:
Context view has a subscription to nuclide-definition-service, and asks for new definitions when the cursor position changes. If the cursor position changes too fr
equently, the definition service can overload the hack service with requests, resulting in errors and sometimes a redbox. This diff addresses some of those issues:

Reviewed By: hansonw

Differential Revision: D3479502

fbshipit-source-id: 175a596b17db69768c19be5cb12be19b8ce7b3ec