Add ModalMultiSelect component
Summary:
This component combines the Modal component in D3485900 and the MultiSelectList from D3485904 to create a multiselect component that uses Atom's modal panel to render its list items.

The intention is to have an API that mimics `<Dropdown />`. In the future, it would be nice to make this support single-selection too.

Reviewed By: jgebhardt

Differential Revision: D3485918

fbshipit-source-id: 1231af1425be4c4de1f4631c02fab6f77818430e