ObjC: remove exceptions from colon/bracket inserters
Summary:
In some cases, Atom's onInsertText event can have an inaccurate location!
This can be the case, for example, if another handler triggers and modifies that line.

In this example, if you type "public:" the trigger that unindents the code interferes with our colon inserter, causing this assertion to redbox.

In practice, we don't actually need to check these, as the calling code already checks based on the text that was inserted.

Reviewed By: zertosh

Differential Revision: D3500786

fbshipit-source-id: da022a8e8f75ca6d7f67c08e655fb00b2057f2a0