Use standardized Section component, allow Section to be controlled.
Summary: In striving for consistency, we should use standard components whenever possible. This replaces the custom toggleable section in the open files list with the standard <Section /> component. So far all use cases of <Section /> have called for uncontrolled usage of the component, so I extended it to accommodate the controlled use case as well. Lastly, since vertical space is especially valuable in the file tree, I added a smaller variant to Section as well.

Reviewed By: ssorallen

Differential Revision: D3481349

fbshipit-source-id: fae63ba6e7c4d3b9e18c93c57a03664d1b2d92a9