Redesign registration API for context view providers
Summary:
Context providers consume the nuclide-context-view service, which previously exposed `registerProvider` and `deregisterProvider` functions. Context providers would have to maintain a handle to the nuclide-context-view service, and call `deregisterProvider` on that handle whenever the context provider was deactivated. This was messy because each context provider had to keep the handle to the context view service up to date which was confusing and required an unnecessary amount of boilerplate.

Now, only `registerProvider` is exposed, and it returns a "registration" to the service, which is a Disposable which can be `dispose()`ed when the context provider is deactivated. The context view service handles the deregister without the context provider having to worry about it.

Reviewed By: hansonw

Differential Revision: D3497552

fbshipit-source-id: 19b3215ab5c30938ac92c4c74a1b7ea70b606814