Ensure that websocket connections are always alive
Summary: We've noticed cases where the heartbeat never reports a problem but the websocket connection is broken - from a UI perspective the connection appears ok but nothing works. This change uses WS's ping/pong mechanism to verify that the websocket is still alive. The client will send the server a "ping", and wait 5 seconds for a "pong". If the "pong" comes back, then it schedules another "ping" to run in 5 seconds. If the "pong" doesn't come back, the websocket is closed. Closing the websocket triggers the mechanism that attempts to reconnect.

Reviewed By: hansonw

Differential Revision: D3507999

fbshipit-source-id: 1c591e023bba7eb07e60cb6f508a64ca6a7494b6