Send pings and forwards pongs only when the socket is open
Summary: In `NuclideSocket`, the timer that calls `ping` is removed when the socket closes. However, sometimes it seems that the socket can be closed but `NuclideSocket` doesn't know it. So we fail the invariant. Other code in `WebSocketTransport` checks for the socket before doing things, instead of using an invariant.

Reviewed By: peterhal

Differential Revision: D3509698

fbshipit-source-id: 24baa844772c9b3da37a87b9e1e5dd8aacbc26cc