Fix tool-bar disposal and support v0 and v1
Summary:
The `tool-bar` package will soon release v1.0.0. In this update, it has dropped `space-pen`, so `addButton` no longer returns a `space-pen` view. However, it does expose the underlaying DOM element the same way - via an `element` property. This diff changes our `tool-bar` usage to always use `element` (when needed), and sets the service API version requirement to `^0 || ^1` (since this is a backwards-compat change).

A lot of packages weren't properly disposing of `tool-bar` related resources. This was most obvious with packages that showed state on `tool-bar` element. If `tool-bar` was disabled, and then re-enabled, the counters would never update. Less obvious was that by subscribing to the cleanup disposable when either the tool-bar or the package was disabled, meant that the other side's disposable still held a reference to the toolbar.

See: https://github.com/suda/tool-bar/issues/141.

Reviewed By: jamesgpearce

Differential Revision: D3513888

fbshipit-source-id: f995d3e9c686876d36924054cf1dd2f88e1650a7