Fix MeasuredComponent attempting to observe null DOM node upon component unmount
Summary: When the ref callback in MeasuredComponent is called on unmount, the node passed in is null. MeasuredComponent now handles that case by first disconnect()ing and then only observing the new node if it is not null. Since _updateNode is called upon unmount, no need to disconnect() in componentWillUnmount()

Reviewed By: ssorallen

Differential Revision: D3503863

fbshipit-source-id: e77edffc255170bef1846ee09fc699679748ba17