Choose last used profile in Connection Dialog
Summary:
Rather than select "Most Recent" by default, select the last used
profile if it still exists. This is more friendly to the file-tree,
which depends on named profiles being used in order to label remote
roots with their profile names.

The data required to enable this feature was already being serialized;
nothing new was added to serialization.

changelog:

* The remote connection dialog now selects your last used profile when
  it's opened rather than "(default)" / "Most Recent"

Reviewed By: matthewwithanm

Differential Revision: D3516845

fbshipit-source-id: 8f15e6175a0f862405bf54db185863a2dc477a40