Current working directory: fix for reloading remote projects
Summary:
When remote projects are reloaded, initially they're not added to Atom's list of directories.
It's only when the remote-connection package finishes loading that this happens.

Unfortunately, this means that the saved CWD appears to be initially invalid. We need to forcibly refresh the CWD when Atom's list of projects changes - this diff implements that.

Reviewed By: matthewwithanm

Differential Revision: D3512254

fbshipit-source-id: fd5ad7eb0984c4df57129ab752cd21aac2c45ef4