Filter handles into processes etc at the root of the health UI
Summary:
THe root component of the health panel should know about 'handles by type' rather than just the rendering down in the panel itself. This is because we might want to surface child process counters in other places.

No change to the logic here, just moving it up to the main controller.

Reviewed By: matthewwithanm

Differential Revision: D3513390

fbshipit-source-id: 9bca052206188b3975db78949d494931ab7814b2