Move the FileTreeNode derived fields calculating logic to a separate class and momoize it
Summary:
FileTreeNode contains a bunch of calculate fields, but itself is immutable. Previously these fields were copied over on a best-effort basis, (too) often falling back to recalculating them. This diff makes things better by:
1) Moving the calculation (derivation) logic out of the FileTreeNode
2) Memoizing the last result

The logic inside the FileTreeNode mostly stayed as it was - it doesn't care about the momoization.
The impact of this is mostly felt when working with large (>2000) file-tree and when have a Working Set enabled. The calculation whether a node belongs or not to a Working Set was somewhat expensive. This diff (along with the one it depends on) help remove the vast majority of redundant recalculations.

Reviewed By: peterhal

Differential Revision: D3505760

fbshipit-source-id: 29ea083754d8f9e9787cecfc335222d8bc3bb573