Buck: store project root, not the project object
Summary:
Caching `BuckProject` is actually bad - for example, if you remove the remote directory and re-add it, the Buck toolbar will be FUBARed until you reload Atom.

This converts the toolbar to only cache the project root. This also makes it easy to consume and display the root path for analytics & UI purposes.

There's a slight perf overhead due to object construction, but all of these actions are in response to user actions so it won't be noticeable.

Reviewed By: matthewwithanm

Differential Revision: D3520844

fbshipit-source-id: 39222f631fc4505e5bf70b56da1d61f45044068c