Adding the UniversalDisposable class
Summary:
The utility class to unite all kind of tear-down logic.
The incentive: we have three ways of performing a tear-down and resource release. Disposables, subscriptions, and plain functions. In addition to these, there's also a bunch of helper classes that help aggregating instances of the same class and classes that handle adaption of one kind to the interface of the other.
This makes little sense. Instead you should use this UniversalDisposable.

It's a drop-in replacement for `Atom`'s/`event-kit`'s `Disposable` and `CompositeDisposable`, but it also implements a logic similar to that of the `RX`'s subscription.
You can construct it with a var-args of subscriptions, disposables and/or plain functions. Or you can `add` to it a var-args of the same.
You can remove previously added var-args.

From now on just use this class. A refactoring of our code is coming up in later diffs.

Reviewed By: matthewwithanm

Differential Revision: D3526551

fbshipit-source-id: 8f1235e4c60243fd64c73a1965ab98367734c21c