Properly unmount the side bar component on deactivate
Summary: The component wasn't being unmounted correctly so React was maintaining a reference to the component, its props, and DOM node, after Nuclide was deactivated. Because `renderPanelSync` is debounced, it's possible that the last `setState` causes it to run after Nuclide has been deactivated. This caused the component to be re-mounted. Now the `panel` variable is nullable to indicate that the package is deactivated.

Reviewed By: ssorallen

Differential Revision: D3548116

fbshipit-source-id: 13a17c97484ea9d02812a546d632115feea206d7