Properly unmount nuclide-type-coverage's status bar tile component on deactivate
Summary: The component wasn't being unmounted correctly so React was maintaining a reference to the component, its props, and DOM node, after Nuclide was deactivated. `add` returns the child subscription, so when `this.subscription.unsubscribe` was called, only the `isActive` subscription was being unsubscribed from. Because the component was now properly unmounted, the still alive `results` subscription was failing an invariant.

Reviewed By: nmote

Differential Revision: D3548177

fbshipit-source-id: a441747bd00f040ff0148e33adb454aa24eb346e