Properly unmount the quick open selection component on deactivate
Summary: The component wasn't being unmounted so React was maintaining a reference to the component, its props, and DOM node, after Nuclide was deactivated. The `_isMounted` checks are necessary because it's possible that after the component is unmounted, for the debounce timer to fire one last time, resulting in an assertion failure.

Reviewed By: jgebhardt

Differential Revision: D3548098

fbshipit-source-id: 25c69280d79ca71f45e50bec125e9092740ffffa