Properly unmount the hhvm toolbar on deactivate
Summary: As the removed comment metions, for some reason `ReactDOM.findDOMNode` returns `null` when the toolbar is hidden. However, React still maintained a referece to the component, its props, and DOM node. This diff moves the unmount to where the toolbar is created, so we can call `ReactDOM.unmountComponentAtNode` with a direct reference to the root node.

Reviewed By: johnislarry

Differential Revision: D3546139

fbshipit-source-id: 21c2679e6d803ef48da51bb4c07d0d4eb8b7290c