Linter adapter extension: invalidateOnClose
Summary:
When the `invalidateOnClose` flag is set to true in a Linter API provider, any diagnostics generated by the file/buffer will be invalidated on close.

This is the most commonly desired behaviour, and the OCaml/Python providers have been changed to use it.

This also adds a check to make sure the editor has not been destroyed by the time lint results come back.

Reviewed By: nmote

Differential Revision: D3067168

fbshipit-source-id: b82ef69e79c9338b54847000f69a4fdbdceeeccc