Use real clock in Flow server startup test
Summary:
This is testing code that pings the Flow server at 1 second intervals. It turns
out that it's basically an accident that this test passes right now, and some
upcoming changes cause it to fail.

It's nontrivial to call `advanceClock` at the right times. So, use the real
clock instead, for this one test.

Reviewed By: zertosh

Differential Revision: D3552955

fbshipit-source-id: bc854d416b9e46d159091c49b62f62dc33f6fc11