Bind shortcuts on nuclide-context-view activation
Summary:
nuclide-context-view is toggled/shown via functions bound to keybindings in `ContextViewManager`. The manager is not initialized in the case that context view consumes no services and isn't consumed by anything, which makes Context View inaccessible. This moves the keybinding to package activation.

While this isn't super impactful now, the activation lifecycle could change in the future and it is probably better practice to bind these in activation.

Reviewed By: jgebhardt

Differential Revision: D3558427

fbshipit-source-id: afa199e0d3c6d18542577f1043c936d1ffb89d2c