Add "Open in main editor" button in Definition Preview
Summary:
This adds a button in the Definition Preview context provider that opens the definition currently being previewed in the main editor. The main editor opens the file with the cursor at the start of the definition. While using context view to write the GraphQL endpoints in www, I found that this feature would be massively helpful.

Cursor is over a class name:
{F61974740}

Clicked the button, class is opened:
{F61974837}

Cursor is over function call:
{F61974925}

Clicked the button, function opened:
{F61974956}

Cursor is over variable:
{F61974993}

Clicked the button, cursor moved to variable declaration:
{F61975000}

Reviewed By: jgebhardt

Differential Revision: D3551317

fbshipit-source-id: 54d4b3f63883138cd1e9cfc9493bccc1c9770cb7