Accumulate tasks from all systems into a single list [1/2]
Summary: This changes the model of nuclide-build. Instead of having a single build system selected at a time, we accumulate the tasks from all of them. This is a prerequisite for this becoming a more generic "task runner," I think: you should be able to use a "Jest task runner" and "RN task runner", for example, without having to switch between the two every time you want to run a task. It will also give us a better story for default behavior (since we can leave it to build systems to determine if they have applicable tasks or not).

Reviewed By: hansonw

Differential Revision: D3562065

fbshipit-source-id: e88a67e2694f4401aed752846930b511d1febc71