Add integration test for react unmounting
Summary: After every integration test deactivates the Nuclide package, this tests for still mounted React roots. It tries to be helpful with the error message by giving you the top components name. This turns out to be very unhelpful if you're using a generic wrapper - that's why the debugging instructions are printed.

Reviewed By: jamesgpearce

Differential Revision: D3561217

fbshipit-source-id: 4dd3846bff20b75c00d4cc2ff6bf593fd5fa1f88