Fix ws closing due to missing pong on messages saturation
Summary:
When on a low-bandwidth/high-latency - but otherwise healthy - connection, it is possible that an over-saturation of received messages causes the pong response to be missed. This then results in the connection getting closed by the ping timer. This diff mitigates this (a bit) by resetting the ping timer when messages are received.

There's a huge flaw in the various Transport classes: They assume that a `close()` is synchronous, and that it's ok to drop messages received between when calling `close()` and when the socket actually closes. The ws socket is closed by calling its own `close()` – this tries to inform the server that the client wants to close, and waits for the response. This can take a longgg time.

Also, in these slow but healthy connections, it is possible for the heartbeat to fail, and then reconnect. Since the ping is already aggressively checking the connection, the heartbeat reconnect is now only allowed if the ws transport is closed.

Reviewed By: johnislarry

Differential Revision: D3576648

fbshipit-source-id: 631541c1f6145200acdbf090676e71e3d433ff74