Compute and cache all info necessary to exec Flow as one unit
Summary:
This makes the various parts easier to reason about and paves the way for the
setting to use the `flow-bin` copy of Flow (#570), since we can have a separate
path to Flow for each root.

One unpleasant side effect of this change is that we are now caching the
pathToFlow setting, when we weren't before. That means that after you update the
setting, it can take up to 30 seconds to take effect. I will explore options to
improve this but I don't think it should block this diff.

This kills a lot of code which I will remove in a separate diff.

Reviewed By: mostafaeweda

Differential Revision: D3555803

fbshipit-source-id: e56cae562ae8f5f522aba568fe867fbba751b80c