Add a "Toggle Toolbar" command for each task runner
Summary: This adds a palette command for each task runner to improve discoverability. If you try to "Toggle X Toolbar" while Y is visible, X will be shown; otherwise it toggles like normal. If there are no tasks for the toolbar you want, it will just show the task runner toolbar and pop up an Atom warning saying there are no tasks for that toolbar. (This should basically never happen.)

Reviewed By: jgebhardt

Differential Revision: D3575641

fbshipit-source-id: d25740c58d0604e9f362a41441d1a43fbf3cbf72