Setting to use the flow-bin binary in the current project
Summary:
As noted in the setting description and in #570, enabling this setting is a
security risk unless you fully trust all of the projects that you open. However,
it can be very useful if you are willing to only open projects you trust. It
can't be the default, but there has been enough demand for such a setting that
it was worth implementing.

Reviewed By: zertosh

Differential Revision: D3590454

fbshipit-source-id: e9b391459ccded64b98d976bf22b43ba28033910