Make Flow outline work with per-project flow-bin binary
Summary:
We want to support Flow outline view outside of Flow root, and even for untitled
files. Obviously, we have to use the system-wide Flow for that. But we also want
to support Flow outline view if the only Flow you have is a local `flow-bin`
installation (assuming you have the setting to use `flow-bin` checked).

See #570

Reviewed By: zertosh

Differential Revision: D3593787

fbshipit-source-id: 8e11b3156e9db379026b66650c904d61ddd7f493