Add priority rendering to Context View
Summary: Context View providers should be rendered with the most frequently useful ones at the top so that the most valuable information is displayed. This is implemented by adding a `priority` member to the Context Provider type so that the array of context providers rendered is ordered by priority (lower priority is rendered higher in the side panel).

Reviewed By: jgebhardt

Differential Revision: D3599393

fbshipit-source-id: 6b577afc5368b6b73cfe18d7bc7207a0fe6a8bfe