Use synchronous compression/decompression for small WebSocket payloads
Summary:
The default compression/decompression protocol in WebSockets uses the async API:
https://github.com/websockets/ws/blob/master/lib/PerMessageDeflate.js#L224
However, this means that we wait for at least one event tick for every RPC message to be processed.

This makes remote project startup appear extremely sluggish - since so much else is going on, the decompress calls can take over 500ms each to process. This means that the user has to wait quite a while for all the files to be reloaded, with a large delay in between (causing a cascading effect).

By decompressing messages synchronously, we avoid having this blockage. As long as the payload is fairly small, the actual compress/decompress is extremely fast and doesn't block for very long.

Video of before and after: https://www.facebook.com/pxlcld/p47Q

A numeric benchmark of reloading 12 remote fbcode files shows a huge win (5 runs each, all ms):

|Before|After|
|----------|----------|
|4508.254|1933.607|
|5338.547|1868.945|
|4975.520|1716.133|
|5102.432|1777.684|
|4810.297|1686.666|
------------------

Average (before): **4947.01**
Average (after): **1796.61**

Reviewed By: zertosh

Differential Revision: D3575294

fbshipit-source-id: 7f27255f16101c53d4dc636828af20eb2138650f