Make ActiveEditorRegistry re-query providers when a new one is added
Summary:
I've noticed recently that on Nuclide restart, outline view reports that the
current grammar is unsupported. I don't know why I didn't see that before, but
anyway it was just coincidentally working properly.

This re-queries providers whenever a new one is added. We could cut down on
re-queries a bit by being smarter here but given that providers are only added
rarely it's probably not a big deal.

Reviewed By: peterhal

Differential Revision: D3606014

fbshipit-source-id: 8ef4c79c436b55b859a6737fe9ad5ae5dc53547f