Related Files: attach command listener to workspace rather than editor
Summary:
I think the integration test is flaky because we send the command before the listener can be attached to the editor.

Instead, let's just attach the command listener to the global workspace and use the active text editor. This simplifies the code and lets us add a menu item for discoverability.

Reviewed By: zertosh

Differential Revision: D3616159

fbshipit-source-id: d605e8b9d1a22e6a59fe44d0830c852bddaba196