Silence missing .git RemoteDirectory log messages
Summary:
Since Atom's Git repository code is still synchronous, it creates many RemoteDirectories for potential `.git` directories in the file system (which fail to subscribe).

This is particularly painful for integration tests, since any non-Hg fixture will be spammed with these error messages. A quick, hacky fix is to silence errors for missing .git directories.

We should remove this when Atom switches their repository API to be fully async.

Reviewed By: zertosh

Differential Revision: D3633863

fbshipit-source-id: 92c9717e285169c4c207deffa1b7f7e7483c19a3